partial backport of changes from signal_views branch
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@2579 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ libkwave/undo/UndoDeleteLabelAction.cpp
@@ -88,9 +88,9 @@ UndoAction *UndoDeleteLabelAction::undo(SignalManager &manager,
 //***************************************************************************
 void UndoDeleteLabelAction::dump(const QString &indent)
 {
-    qDebug("%sundo delete label, %u, '%s'",
+    qDebug("%sundo delete label, %lu, '%s'",
            indent.toLocal8Bit().data(),
-           m_label.pos(),
+           static_cast<unsigned long int>(m_label.pos()),
            m_label.name().toLocal8Bit().data()
           );
 }
